refactor(app): extract the query-running wiring
The Run button, re-running or inserting a history entry, the saved-query
list and its context menu, running into the right split pane, Explain
and Format move to `wire::query`.

The runners themselves stay in `main` — they are built from the driver
slot — and reach the module through `AppFns`, which gains `run_stream`
and `sync_editor`. `sync_editor` picks up an explicit `Rc<dyn Fn(usize)>`
type for that.
